# backend-java/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Instant Quote Service (quote-extui-service)",
    "description": "Instant Quote backend: form defaults, master data, city and zip code suggestions, and quote submission.\n\nEvery code field holds an md_codedetail.cdcode exactly as /api/v1/masterdata/codes returned it, for example TPMA and not AIR. Codes are case sensitive. Build the dropdowns from that endpoint rather than hardcoding values.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/profiledata/defaults": {
      "get": {
        "tags": [
          "ProfileData"
        ],
        "summary": "Values the quote form starts with",
        "description": "Every field except version is a cdcode, and each one matches a field name on the quote request body, so the response can be copied into the form state as it is. No parameters.",
        "operationId": "defaults",
        "responses": {
          "200": {
            "description": "The current defaults.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteDefaults"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. The body never carries internal detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/suggestions": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Suggest cities and zip codes as the user types",
        "description": "Calls Amazon Location Service (Places API v2 Autocomplete). Returns at most 8 suggestions, cities and zip codes only.",
        "operationId": "suggest",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "What the user typed. 2 to 100 characters.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Chicago"
          },
          {
            "name": "countryCode",
            "in": "query",
            "description": "Restrict to one ISO country. Case insensitive.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "US"
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions, best match first.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponsePlaceSuggestion"
                }
              }
            }
          },
          "400": {
            "description": "A parameter broke a rule.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Amazon Location Service failed or the API key is not set.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. The body never carries internal detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/masterdata/codes": {
      "get": {
        "tags": [
          "MasterData"
        ],
        "summary": "List code categories and their active values",
        "description": "Build every dropdown from this; several categories come back per call. A category that does not exist is left out rather than returning 404, so compare what you asked for against what came back to catch a typo. A category with no active value returns an empty codes array.",
        "operationId": "listCodes",
        "parameters": [
          {
            "name": "cmcode",
            "in": "query",
            "description": "Code category to return, for example CGT. Repeatable, and a single value may hold a comma separated list, so cmcode=CGT,PKT&cmcode=TPM works. Omit entirely to get every active category. Each token must match ^[A-Z0-9]{1,3}$ and at most 20 distinct categories are accepted.",
            "schema": {
              "type": "string"
            },
            "example": "CGT,PKT"
          }
        ],
        "responses": {
          "200": {
            "description": "Categories sorted by cmcode.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseCodeMaster"
                }
              }
            }
          },
          "400": {
            "description": "A cmcode broke the pattern, or more than 20 were requested.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. The body never carries internal detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masterdata/airports": {
      "get": {
        "tags": [
          "MasterData"
        ],
        "summary": "Search airports by IATA code, city or name",
        "description": "Best match first: exact IATA code, then IATA prefix, then city prefix. Equal matches come back largest airport first. About 9,000 airports are loaded, all with an IATA code.",
        "operationId": "searchAirports",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Free text. At least 2 characters, at most 60. Letters, digits, spaces, apostrophes, hyphens and periods only. Matches the IATA code, the city and the airport name, case insensitive. Omit for no text filter.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "chi"
          },
          {
            "name": "countryCode",
            "in": "query",
            "description": "Restrict to one ISO country. Case insensitive.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "US"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum rows. Defaults to 20, maximum 100.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "Matching airports, best match first.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseAirport"
                }
              }
            }
          },
          "400": {
            "description": "A parameter broke a rule, for example a one character query.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. The body never carries internal detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "RFC 9457 problem detail. Field level failures add errors.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "The request body has 2 validation error(s)."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/quote-requests"
          },
          "errors": {
            "type": "array",
            "description": "One entry per rejected field. Absent when no field was at fault.",
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            }
          }
        }
      },
      "ApiFieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "lineItems[0].packageType"
          },
          "message": {
            "type": "string",
            "example": "'BOX' is not an active code. Expected a cdcode from category PKT."
          }
        }
      },
...
      "QuoteDefaults": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32",
            "description": "Goes up whenever a value below changes. Compare it to decide whether a cached copy is stale.",
            "example": 1
          },
          "mode": {
            "type": "string",
            "description": "Category TPM.",
            "example": "TPMA"
          },
          "cargoType": {
            "type": "string",
            "description": "Category CGT.",
            "example": "CGTPNP"
          },
          "ratingType": {
            "type": "string",
            "description": "Category RTT.",
            "example": "RTTPU"
          },
          "pickupType": {
            "type": "string",
            "description": "Category PDT.",
            "example": "PDTPO"
          },
          "deliveryType": {
            "type": "string",
            "description": "Category PDT.",
            "example": "PDTPO"
          },
          "weightUom": {
            "type": "string",
            "description": "Category WUM.",
            "example": "WUMKG"
          },
          "dimensionUom": {
            "type": "string",
            "description": "Category DUM.",
            "example": "DUMCM"
          },
          "volumeUom": {
            "type": "string",
            "description": "Category VUM.",
            "example": "VUMCBM"
          }
        }
      },
      "ListResponsePlaceSuggestion": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestion"
            }
          }
        }
      },
      "PlaceSuggestion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Amazon Location place id."
          },
          "city": {
            "type": "string",
            "example": "Chicago"
          },
          "state": {
            "type": "string",
            "example": "IL"
          },
          "zipCode": {
            "type": "string",
            "example": "60601"
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "countryCode": {
            "type": "string",
            "example": "US"
          },
          "label": {
            "type": "string",
            "description": "The full address line from Amazon Location.",
            "example": "60601, Chicago, IL, United States"
          }
        }
      },
...
      "CodeDetail": {
        "type": "object",
        "properties": {
          "cdcode": {
            "type": "string",
            "description": "Send this value back on a quote request.",
            "example": "PKTBOX"
          },
          "description": {
            "type": "string",
            "description": "Label to show the user.",
            "example": "BOX"
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "description": "Display order. Null sorts last.",
            "example": 20
          }
        }
      },
      "CodeMaster": {
        "type": "object",
        "properties": {
          "cmcode": {
            "type": "string",
            "example": "PKT"
          },
          "description": {
            "type": "string",
            "example": "Package Type"
          },
          "codes": {
            "type": "array",
            "description": "Active values, sorted by sequence then cdcode. May be empty.",
            "items": {
              "$ref": "#/components/schemas/CodeDetail"
            }
          }
        }
      },
      "ListResponseCodeMaster": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeMaster"
            }
          }
        }
      },
      "Airport": {
        "type": "object",
        "properties": {
          "iataCode": {
            "type": "string",
            "description": "Send this on a quote request as the port code.",
            "example": "ORD"
          },
          "icaoCode": {
            "type": "string",
            "example": "KORD"
          },
          "name": {
            "type": "string",
            "example": "Chicago O'Hare International Airport"
          },
          "municipality": {
            "type": "string",
            "description": "City the airport serves.",
            "example": "Chicago"
          },
          "isoRegion": {
            "type": "string",
            "description": "ISO 3166-2 subdivision.",
            "example": "US-IL"
          },
          "isoCountry": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2.",
            "example": "US"
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "example": 41.9786
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "example": -87.9048
          }
        }
      },
      "ListResponseAirport": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Airport"
            }
          }
        }
      }
    }
  }
}